# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "description":"**🚀 Getting Started Guide:**\n\n1. **Register**: Use `/api/users/signup` to create a new account (Customer or Seller)\n2. **Login**: Use `/api/users/signin` with your credentials to get a JWT token\n3. **Authorize**: Click the 🔒 **Authorize** button above and enter: `Bearer <your-jwt-token>`\n4. **Explore**: Now you can access all protected endpoints!\n\n---\n\n**Complete E-commerce API** with role-based access control, product management, order processing, cart functionality, and social features.\n\n**User Roles:**\n- **Customer**: Can browse products, manage cart, place orders, like items\n- **Seller**: Can manage own products, view related orders, access analytics\n- **Admin**: Full system access and user management",
    "title": "E-commerce API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "JWT": {
      "in": "header",
      "name": "Authorization",
      "type": "apiKey",
      "description": "Enter JWT token in format: Bearer <token>"
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get All Products",
        "description": "Get all active products (accessible to all authenticated users)",
        "security": [{ "JWT": [] }],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "count": { "type": "number" },
                "products": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/Product" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Add Product",
        "description": "Add a new product (Sellers and Admins only)",
        "security": [{ "JWT": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "message": { "type": "string" },
                "product": { "$ref": "#/definitions/Product" }
              }
            }
          },
          "400": {
            "description": "Invalid product data"
          },
          "403": {
            "description": "Only sellers and admins can add products"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/cartItems": {
      "get": {
        "tags": ["Cart Items"],
        "summary": "Get Cart Items",
        "description": "Get current user's cart items with product details",
        "security": [{ "JWT": [] }],
        "responses": {
          "200": {
            "description": "Cart items retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "count": { "type": "number" },
                "totalAmount": { "type": "number" },
                "cartItems": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/CartItem" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Cart Items"],
        "summary": "Add to Cart",
        "description": "Add product to cart or update quantity if already exists",
        "security": [{ "JWT": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Cart item data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CartItemCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Item added to cart successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "message": { "type": "string" },
                "cartItem": { "$ref": "#/definitions/CartItem" }
              }
            }
          },
          "400": {
            "description": "Invalid product ID or quantity"
          },
          "404": {
            "description": "Product not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Cart Items"],
        "summary": "Clear Cart",
        "description": "Remove all items from user's cart",
        "security": [{ "JWT": [] }],
        "responses": {
          "200": {
            "description": "Cart cleared successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "message": { "type": "string" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/likes": {
      "get": {
        "tags": ["Likes"],
        "summary": "Get Likes",
        "description": "Get likes for a specific entity (product or category)",
        "security": [{ "JWT": [] }],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": ["Product", "Category"],
            "description": "Type of entity to get likes for"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "ID of the entity"
          }
        ],
        "responses": {
          "200": {
            "description": "Likes retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "likesCount": { "type": "number" },
                "userHasLiked": { "type": "boolean" },
                "likes": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/Like" }
                }
              }
            }
          },
          "400": {
            "description": "Invalid type or ID parameter"
          },
          "404": {
            "description": "Entity not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Likes"],
        "summary": "Toggle Like",
        "description": "Like or unlike a product or category",
        "security": [{ "JWT": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "ID of the entity to like"
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": ["Product", "Category"],
            "description": "Type of entity to like"
          }
        ],
        "responses": {
          "200": {
            "description": "Like toggled successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "message": { "type": "string" },
                "isLiked": { "type": "boolean" },
                "likesCount": { "type": "number" }
              }
            }
          },
          "400": {
            "description": "Invalid type or ID parameter"
          },
          "404": {
            "description": "Entity not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create Order",
        "description": "Place a new order (Customers only)",
        "security": [{ "JWT": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Order data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order placed successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "message": { "type": "string" },
                "order": { "$ref": "#/definitions/Order" }
              }
            }
          },
          "400": {
            "description": "Invalid order data or insufficient stock"
          },
          "403": {
            "description": "Only customers can place orders"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Product": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "name": { "type": "string" },
        "description": { "type": "string" },
        "price": { "type": "number" },
        "inStock": { "type": "number" },
        "categories": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": { "type": "string" },
              "name": { "type": "string" },
              "description": { "type": "string" }
            }
          }
        },
        "seller": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "name": { "type": "string" },
            "email": { "type": "string" },
            "type": { "type": "string" }
          }
        },
        "ratings": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "userID": {
                "type": "object",
                "properties": {
                  "_id": { "type": "string" },
                  "name": { "type": "string" }
                }
              },
              "rating": { "type": "number", "minimum": 1, "maximum": 5 }
            }
          }
        },
        "reviews": {
          "type": "array",
          "items": { "type": "string" }
        },
        "isActive": { "type": "boolean" },
        "createdAt": { "type": "string", "format": "date-time" },
        "updatedAt": { "type": "string", "format": "date-time" }
      }
    },
    "ProductCreate": {
      "type": "object",
      "required": ["name", "description", "price", "categories"],
      "properties": {
        "name": { 
          "type": "string",
          "example": "iPhone 14 Pro"
        },
        "description": { 
          "type": "string",
          "example": "Latest iPhone with advanced camera system"
        },
        "price": { 
          "type": "number",
          "example": 999.99
        },
        "inStock": { 
          "type": "number", 
          "default": 0,
          "example": 10
        },
        "categories": { 
          "type": "array",
          "items": { "type": "string" },
          "example": ["Electronics", "Mobile Phones"]
        }
      }
    },
...
    "CartItem": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "userID": { "type": "string" },
        "productID": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "name": { "type": "string" },
            "description": { "type": "string" },
            "price": { "type": "number" },
            "inStock": { "type": "number" },
            "seller": {
              "type": "object",
              "properties": {
                "_id": { "type": "string" },
                "name": { "type": "string" },
                "email": { "type": "string" }
              }
            }
          }
        },
        "quantity": { "type": "number" },
        "totalPrice": { "type": "number" },
        "createdAt": { "type": "string", "format": "date-time" },
        "updatedAt": { "type": "string", "format": "date-time" }
      }
    },
    "CartItemCreate": {
      "type": "object",
      "required": ["productID", "quantity"],
      "properties": {
        "productID": { 
          "type": "string",
          "example": "60d5ecb74c5b1c2d8c8b4567"
        },
        "quantity": { 
          "type": "number", 
          "minimum": 1,
          "example": 2
        }
      }
    },
    "Like": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "userID": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "name": { "type": "string" },
            "email": { "type": "string" }
          }
        },
        "likeableType": {
          "type": "string",
          "enum": ["Product", "Category"]
        },
        "likeableID": { "type": "string" },
        "entityDetails": {
          "type": "object",
          "description": "Populated details of the liked entity (Product or Category)"
        },
        "createdAt": { "type": "string", "format": "date-time" }
      }
    },
    "OrderCreate": {
      "type": "object",
      "required": ["items", "shippingAddress"],
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "required": ["productID", "quantity"],
            "properties": {
              "productID": { 
                "type": "string",
                "example": "60d5ecb74c5b1c2d8c8b4567"
              },
              "quantity": { 
                "type": "number", 
                "minimum": 1,
                "example": 2
              }
            }
          }
        },
        "shippingAddress": {
          "type": "object",
          "required": ["street", "city", "state", "zipCode"],
          "properties": {
            "street": { 
              "type": "string",
              "example": "123 Main Street"
            },
            "city": { 
              "type": "string",
              "example": "Mumbai"
            },
            "state": { 
              "type": "string",
              "example": "Maharashtra"
            },
            "zipCode": { 
              "type": "string",
              "example": "400001"
            },
            "country": { 
              "type": "string", 
              "default": "India",
              "example": "India"
            }
          }
        },
        "paymentMethod": {
          "type": "string",
          "enum": ["Credit Card", "Debit Card", "PayPal", "Cash on Delivery", "UPI"],
          "default": "Cash on Delivery",
          "example": "Cash on Delivery"
        },
        "orderNotes": { 
          "type": "string",
          "example": "Please deliver after 6 PM"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "customer": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "name": { "type": "string" },
            "email": { "type": "string" }
          }
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "product": {
                "type": "object",
                "properties": {
                  "_id": { "type": "string" },
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" }
                }
              },
              "seller": {
                "type": "object",
                "properties": {
                  "_id": { "type": "string" },
                  "name": { "type": "string" },
                  "email": { "type": "string" }
                }
              },
              "quantity": { "type": "number" },
              "price": { "type": "number" },
              "totalPrice": { "type": "number" }
            }
          }
        },
        "shippingAddress": {
          "type": "object",
          "properties": {
            "street": { "type": "string" },
            "city": { "type": "string" },
            "state": { "type": "string" },
            "zipCode": { "type": "string" },
            "country": { "type": "string" }
          }
        },
        "paymentMethod": { "type": "string" },
        "paymentStatus": { 
          "type": "string",
          "enum": ["Pending", "Paid", "Failed", "Refunded"]
        },
        "orderStatus": { 
          "type": "string",
          "enum": ["Pending", "Confirmed", "Processing", "Shipped", "Delivered", "Cancelled"]
        },
        "totalAmount": { "type": "number" },
        "discountAmount": { "type": "number" },
        "taxAmount": { "type": "number" },
        "shippingCost": { "type": "number" },
        "finalAmount": { "type": "number" },
        "orderNotes": { "type": "string" },
        "trackingNumber": { "type": "string" },
        "estimatedDelivery": { "type": "string", "format": "date-time" },
        "deliveredAt": { "type": "string", "format": "date-time" },
        "cancelledAt": { "type": "string", "format": "date-time" },
        "cancelReason": { "type": "string" },
        "createdAt": { "type": "string", "format": "date-time" },
        "updatedAt": { "type": "string", "format": "date-time" }
      }
    }
  }
}